requirements.txt
django-cors-headers==3.4.0
Django==2.2.4
django-filter==2.3.0
django-storages==1.7.1
djangorestframework==3.9.0
psycopg2==2.8.3
